Refactor formation demo into dedicated modules
