+++ src/formation_flight_pid_control/__init__.py
@@ -1,2 +1,3 @@
 """Formation Flight PID Control package."""
+
 __version__ = "0.1.0"
